backend/pyproject.toml
[project]
name = "iot-api-store-sensor"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.136.1",
    "google-cloud-firestore>=2.27.0",
    "pydantic>=2.13.4",
    "uvicorn>=0.46.0",
]

[dependency-groups]
dev = [
    "httpx>=0.28.1",
    "pytest>=9.0.3",
    "pytest-mock>=3.15.1",
]
